Completely overhauled world generation architecture
Deprecated the old world generation script with a modular generation system.

It functions in stages. Starting by generating a base of land mass (land and water) then applying modifications to that world.

Each routine is a different stage object. It's easy to add more or modify stages.

Old world generation script still exists for now, and i created a wrapper for the new system so that the old script can be ran as a stage in the new system. 

Old system constants will be moved to a GenerationCompanion object.

Debug
====
When in debug mode, Numpad plus is bound to increase the 3d height of the camera, so you can get a better overview of the world.

F3 also adds the test world generation configuration window to the game.