Fixed errors from valid java code blocks
SSVM would throw non fatal errors for valid java code injection blocks.

This was fixed by accepting a class body equiv pattern instead of a statement block.


-----
ALSO
this removes the unused java antlr recogniser. the sms recogniser can function without it, just it's tokens and data.

This should help github in showing 80% of the project being java.

----
ALSO
This changes the state machine script extension to sms, in the hopes that github will no longer recognise it as as scheme file.